Refactor sequence_parallel.py and model.py to ensure tensor dtype consistency for cosine and sine calculations in rotary position embeddings.

+++ wan/distributed/sequence_parallel.py
@@ -124,8 +124,8 @@ def sp_dit_forward(
             s_per_rank = s
             freqs_i_rank = freqs_i[(sp_rank * s_per_rank):((sp_rank + 1) * s_per_rank), :, :]
             cos, sin = torch.chunk(torch.view_as_real(freqs_i_rank.to(torch.complex64)), 2, dim=-1)
-            cos = cos.unsqueeze(0).expand(-1, -1, -1, -1, 2).flatten(-2)
-            sin = sin.unsqueeze(0).expand(-1, -1, -1, -1, 2).flatten(-2)
+            cos = cos.unsqueeze(0).expand(-1, -1, -1, -1, 2).flatten(-2).to(x.dtype)
+            sin = sin.unsqueeze(0).expand(-1, -1, -1, -1, 2).flatten(-2).to(x.dtype)
             freqs_i_rank = (cos, sin)
             freqs_list.append(freqs_i_rank)
         self.freqs_list = freqs_list
